Rename WebAuthJwtFilter
