--- .stellar/identity/anchor-wallet.toml
@@ -1,1 +1,0 @@
-secret_key = "SBHTWEF5U7FK53FLGDMBQYGXRUJ24VBM3M6VDXCHRIGCRG3Z64PH45LW"
