--- .stellar/identity/anchor.toml
@@ -1,1 +1,0 @@
-seed_phrase = "sock prefer gift figure retire love ship tooth sound special onion siege banana follow cry soup access torch base lava crumble fury guitar flip"
