+++ core/src/test/kotlin/org/stellar/anchor/filter/WebAuthJwtFilterTest.kt
@@ -8,8 +8,11 @@ import jakarta.servlet.ServletResponse
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.apache.hc.core5.http.HttpStatus
-import org.junit.jupiter.api.*
 import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Order
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import org.stellar.anchor.TestHelper.Companion.createSep10Jwt
@@ -19,8 +22,8 @@ import org.stellar.anchor.auth.WebAuthJwt
 import org.stellar.anchor.config.AppConfig
 import org.stellar.anchor.config.CustodySecretConfig
 import org.stellar.anchor.config.SecretConfig
-import org.stellar.anchor.filter.Sep10JwtFilter.APPLICATION_JSON_VALUE
-import org.stellar.anchor.filter.Sep10JwtFilter.JWT_TOKEN
+import org.stellar.anchor.filter.WebAuthJwtFilter.APPLICATION_JSON_VALUE
+import org.stellar.anchor.filter.WebAuthJwtFilter.JWT_TOKEN
 import org.stellar.anchor.setupMock
 
 @Order(85)
@@ -33,7 +36,7 @@ internal class WebAuthJwtFilterTest {
   private lateinit var secretConfig: SecretConfig
   private lateinit var custodySecretConfig: CustodySecretConfig
   private lateinit var jwtService: JwtService
-  private lateinit var sep10TokenFilter: Sep10JwtFilter
+  private lateinit var webAuthJwtFilter: WebAuthJwtFilter
   private lateinit var request: HttpServletRequest
   private lateinit var response: HttpServletResponse
   private lateinit var mockFilterChain: FilterChain
@@ -45,7 +48,7 @@ internal class WebAuthJwtFilterTest {
     this.custodySecretConfig = mockk(relaxed = true)
     secretConfig.setupMock()
     this.jwtService = JwtService(secretConfig, custodySecretConfig)
-    this.sep10TokenFilter = Sep10JwtFilter(jwtService)
+    this.webAuthJwtFilter = WebAuthJwtFilter(jwtService)
     this.request = mockk(relaxed = true)
     this.response = mockk(relaxed = true)
     this.mockFilterChain = mockk(relaxed = true)
@@ -57,11 +60,11 @@ internal class WebAuthJwtFilterTest {
     val mockServletResponse = mockk<ServletResponse>(relaxed = true)
 
     assertThrows<ServletException> {
-      sep10TokenFilter.doFilter(mockServletRequest, mockServletResponse, mockFilterChain)
+      webAuthJwtFilter.doFilter(mockServletRequest, mockServletResponse, mockFilterChain)
     }
 
     assertThrows<ServletException> {
-      sep10TokenFilter.doFilter(request, mockServletResponse, mockFilterChain)
+      webAuthJwtFilter.doFilter(request, mockServletResponse, mockFilterChain)
     }
   }
 
@@ -69,7 +72,7 @@ internal class WebAuthJwtFilterTest {
   fun `test OPTIONS method works fine without auth header`() {
     every { request.method } returns "OPTIONS"
 
-    sep10TokenFilter.doFilter(request, response, mockFilterChain)
+    webAuthJwtFilter.doFilter(request, response, mockFilterChain)
 
     verify { mockFilterChain.doFilter(request, response) }
   }
@@ -80,7 +83,7 @@ internal class WebAuthJwtFilterTest {
     every { request.method } returns method
     every { request.getHeader("Authorization") } returns null
 
-    sep10TokenFilter.doFilter(request, response, mockFilterChain)
+    webAuthJwtFilter.doFilter(request, response, mockFilterChain)
     verify(exactly = 1) {
       response.setStatus(HttpStatus.SC_FORBIDDEN)
       response.contentType = APPLICATION_JSON_VALUE
@@ -94,7 +97,7 @@ internal class WebAuthJwtFilterTest {
     every { request.method } returns method
     every { request.getHeader("Authorization") } returns ""
 
-    sep10TokenFilter.doFilter(request, response, mockFilterChain)
+    webAuthJwtFilter.doFilter(request, response, mockFilterChain)
 
     verify(exactly = 1) {
       response.setStatus(HttpStatus.SC_FORBIDDEN)
@@ -108,7 +111,7 @@ internal class WebAuthJwtFilterTest {
     every { request.method } returns method
     every { request.getHeader("Authorization") } returns "Bearer123"
 
-    sep10TokenFilter.doFilter(request, response, mockFilterChain)
+    webAuthJwtFilter.doFilter(request, response, mockFilterChain)
 
     verify(exactly = 1) {
       response.setStatus(HttpStatus.SC_FORBIDDEN)
@@ -120,7 +123,7 @@ internal class WebAuthJwtFilterTest {
   @ValueSource(strings = ["GET", "PUT", "POST", "DELETE"])
   fun `make sure check() exception returns FORBIDDEN and does not cause 500`(method: String) {
     every { request.method } returns method
-    val mockFilter = spyk(sep10TokenFilter)
+    val mockFilter = spyk(webAuthJwtFilter)
     every { mockFilter.check(any(), any(), any()) } answers { throw Exception("Not validate") }
 
     mockFilter.doFilter(request, response, mockFilterChain)
@@ -139,7 +142,7 @@ internal class WebAuthJwtFilterTest {
     every { request.method } returns method
     val mockJwtService = spyk(jwtService)
     every { mockJwtService.decode(any(), AbstractJwt::class.java) } returns null
-    val filter = Sep10JwtFilter(mockJwtService)
+    val filter = WebAuthJwtFilter(mockJwtService)
 
     filter.doFilter(request, response, mockFilterChain)
 
@@ -158,7 +161,7 @@ internal class WebAuthJwtFilterTest {
 
     val jwtToken = jwtService.encode(createSep10Jwt(PUBLIC_KEY, null, "stellar.org"))
     every { request.getHeader("Authorization") } returns "Bearer $jwtToken"
-    sep10TokenFilter.doFilter(request, response, mockFilterChain)
+    webAuthJwtFilter.doFilter(request, response, mockFilterChain)
 
     verify { mockFilterChain.doFilter(request, response) }
     verify(exactly = 1) { request.setAttribute(JWT_TOKEN, any()) }
